+++ src/__stories__/useRefMounted.story.tsx
@@ -5,8 +5,15 @@ import ShowDocs from './util/ShowDocs';
 
 const Demo = () => {
   const refMounted = useRefMounted();
+
   useRaf();
-  return <div>is mounted: {refMounted.current ? '👍' : '👎'}</div>;
+  return (
+    <div>
+      <h3>**DEPRECATED**</h3>
+      <h4>This method is obsolete, use `useMountedState` instead.</h4>
+      <span>is mounted: {refMounted.current ? '👍' : '👎'}</span>
+    </div>
+  );
 };
 
 storiesOf('Lifecycle|useRefMounted', module)