+++ src/index.ts
@@ -57,6 +57,9 @@ export { default as usePrevious } from './usePrevious';
 export { default as usePromise } from './usePromise';
 export { default as useRaf } from './useRaf';
 export { default as useRafLoop } from './useRafLoop';
+/**
+ * @deprecated This hook is obsolete, use `useMountedState` instead
+ */
 export { default as useRefMounted } from './useRefMounted';
 export { default as useScroll } from './useScroll';
 export { default as useScrolling } from './useScrolling';
@@ -76,6 +79,7 @@ export { default as useTween } from './useTween';
 export { default as useUnmount } from './useUnmount';
 export { default as useUpdate } from './useUpdate';
 export { default as useUpdateEffect } from './useUpdateEffect';
+export { default as useUpsert } from './useUpsert';
 export { default as useVideo } from './useVideo';
 export { useWait, Waiter } from './useWait';
 export { default as useWindowScroll } from './useWindowScroll';