+++ src/usePromise.ts
@@ -1,22 +1,18 @@
 import { useCallback } from 'react';
-import useRefMounted from './useRefMounted';
+import useMountedState from './useMountedState';
 
 export type UsePromise = () => <T>(promise: Promise<T>) => Promise<T>;
 
 const usePromise: UsePromise = () => {
-  const refMounted = useRefMounted();
+  const isMounted = useMountedState();
   return useCallback(
     (promise: Promise<any>) =>
       new Promise<any>((resolve, reject) => {
         const onValue = value => {
-          if (refMounted.current) {
-            resolve(value);
-          }
+          isMounted() && resolve(value);
         };
         const onError = error => {
-          if (refMounted.current) {
-            reject(error);
-          }
+          isMounted() && reject(error);
         };
         promise.then(onValue, onError);
       }),