+++ src/useRefMounted.ts
@@ -1,5 +1,8 @@
 import { RefObject, useEffect, useRef } from 'react';
 
+/**
+ * @deprecated This hook is obsolete, use `useMountedState` instead
+ */
 const useRefMounted = (): RefObject<boolean> => {
   const refMounted = useRef<boolean>(false);
 
@@ -14,4 +17,7 @@ const useRefMounted = (): RefObject<boolean> => {
   return refMounted;
 };
 
+/**
+ * @deprecated This hook is obsolete, use `useMountedState` instead
+ */
 export default useRefMounted;